+++ src/validators/posts.rs
@@ -5,7 +5,7 @@ use crate::errors::ValidationError;
 pub const ATTACHMENTS_MAX_NUM: usize = 15;
 pub const MENTIONS_MAX_NUM: usize = 50;
 pub const LINKS_MAX_NUM: usize = 10;
-pub const EMOJIS_MAX_NUM: usize = 20;
+pub const EMOJIS_MAX_NUM: usize = 50;
 
 pub const CONTENT_MAX_SIZE: usize = 100000;
 const CONTENT_ALLOWED_TAGS: [&str; 8] = [